Enable swimming on the singular fish
