+++ apps/typegpu-docs/src/content/examples/rendering/singular-fish/tgsl-helpers.ts
@@ -99,3 +99,43 @@ export const distanceVectorFromLine = tgpu['~unstable']
     return std.sub(point, closestPoint);
   })
   .$name('distance vector from line');
+
+const ApplySinWaveReturnSchema = d.struct({
+  position: d.vec3f,
+  normal: d.vec3f,
+});
+
+export const applySinWave = tgpu['~unstable']
+  .fn([d.u32, d.vec3f, d.vec3f], ApplySinWaveReturnSchema)
+  .does((time, position, normal) => {
+    // z += sin(time / 100 + x) / 5
+    const timeFactor = d.f32(time) / 100;
+
+    const positionModification = d.vec3f(
+      0,
+      0,
+      std.sin(timeFactor + position.x) / 5,
+    );
+
+    const modelNormal = normal;
+    const normalXZ = d.vec3f(modelNormal.x, 0, modelNormal.z);
+
+    const coeff = std.cos(timeFactor + position.x) / 5;
+    const newOX = std.normalize(d.vec3f(1, 0, coeff));
+    const newOZ = d.vec3f(-newOX.z, 0, newOX.x);
+    const newNormalXZ = std.add(
+      std.mul(newOX, d.vec3f(normalXZ.x, 0, 0)),
+      std.mul(newOZ, d.vec3f(0, 0, normalXZ.z)),
+    );
+
+    const wavedNormal = std.normalize(
+      d.vec3f(newNormalXZ.x, modelNormal.y, newNormalXZ.z),
+    );
+
+    const wavedPosition = std.add(position, positionModification);
+
+    return ApplySinWaveReturnSchema({
+      position: wavedPosition,
+      normal: wavedNormal,
+    });
+  });
